Added other personal info fields and consolidated
For now the only truly distinguishing type of FormInput is for textareas, since it's a different HTML element. I do eventually intend to have validation, though, and that's going to require some more distinguishing so I can include patterns and such things as that. Of course there will also be appropriate CSS and a11y features so the user is made aware of any non-valid input. That said, validity here is mostly irrelevant since there's no true submission, but I still think this could be helpful to the user.